stm32h7-startup: move pre_init into assembly.
Our pre-init routine contains things that need to happen before anything
touches RAM. This is not reliably possible from a Rust function, since
the compiler may elect to set up a stack frame (and, in our case,
definitely does so today).

This commit moves it into straight assembly language, ensuring that we
have control over the use of RAM. As a pleasant side effect, this also
makes it really hard to accidentally put Rust code before it.

This is not sufficient, by itself, to eliminate use of RAM early in
boot, since the ancient version of cortex-m-rt that we're using has a
bug here. But I'll tackle that in a follow-on commit.