Upgrade cortex-m-rt and PACs to ~2022
I noticed that our cortex-m-rt version was generating potentially
invalid code in the Reset function -- setting up and tearing down a
stack frame, before we've turned on RAM! This bug got fixed in the
upstream cortex-m-rt crate, but we haven't upgraded in ... quite a
while.

This upgrades to a compromise point -- 0.15 of the stm32 pacs. Upgrading
farther than that causes massive API incompatibilities, since it looks
like they've comprehensively changed the svd2rust generated code. Rather
than touch literally every driver in the system, I've settled on 0.15
and left 0.16 as a trap for the future.

As you can see, this still required a fair amount of churn, mostly due
to renamed enum variants.