Revert "stm32xx-i2c: remove soft timeout, LostInterrupt"
This reverts commit e199470557c39d3accc39e346e907d7a5ef7f905, plus a
small change to adapt to the NotificationBits API, which changed since
this commit was originally applied.

This adds a known-flaky system load detector that crashes the i2c
driver. We're doing this because otherwise the i2c driver still contains
bugs that result in unpredictable hangs, and hangs are currently
believed to be worse than crashes.

I'm really hoping to de-revert this in the very near future.