Add `grapefruit/app-mini.toml` (#2264)
While debugging #2236, I wanted a tiny image that produced compact
system dumps – they're faster to load in a debug build of Humility, and
we'll eventually want to commit one to the Humility test suite.

This image produces a 421KiB dump, versus 7 MiB for a full Cosmo image.
It's just enough to blink an LED so that you know it's been flashed and
is running.