Generate the list of boards to build dynamically (#2263)
Before this PR, we used the `build-boards.yml` reusable workflow to list
which boards were supposed to be built, and then invoked the
`build-one.yml` reusable workflow for each of them. Apart from the extra
indirection of two reusable workflows, this meant every time a new app
is added we had to remember add it to the workflow.

This PR fixes that by merging the two reusable workflows into one, and
generating the build matrix dynamically. There is now a job that runs
before everything else, which clones the repo and runs the
`build/gha-build-boards-matrix.py` script. The script scans the
repository for `app.toml`s and generates the build matrix in JSON
format. The build jobs then depend on the scheduler job, and set their
build matrix to that job's output.

To support building a subset of boards (during releases) the script
optionally accepts a list of paths, and it will look for `app.toml`s
only in them (for example `build/gha-build-boards-matrix.py app/sidecar
app/gimlet`). You can play around with the script locally, it doesn't
require any credentials nor dependencies (except for git and a recent
Python).

The reason why I opted to write the matrix generator in a scripting
language rather than Rust (as part of xtask) is the speed of the
scheduler job. With the current setup, the job in total takes 4 seconds
to run, while running xtask with a full cache would take around a
minute. My hunch is that it's better to deal with a scripting language
rather than making every CI run a minute slower. I chose Python as the
scripting language mostly because it has a TOML parser in its standard
library, and thus doesn't require any dependency.