Fix revision ID pins on Cosmo and Grapefruit (#2269)
Contrary to the comment that _someone_ wrote, these pins are AF on reset
and have to be switched to input mode. We didn't notice this on rev A,
which expected all 0s for the revision.

We'll want this on the `cosmo-b` bringup branch, but it's also wrong on
`master` for rev A, so let's fix it here.

Closes #2268