Tag releases with Brussels (#2270)
The first phase of the SP and RoT release process is to determine the
version number, tag the latest commit with that version number, and push
it to kick off the build. To start fully automating the SP release
process, this PR replaces this phase with a single GitHub Actions
workflow that can be started through the UI:

<img width="933" height="477" alt="image"
src="https://github.com/user-attachments/assets/2fb0b680-b4ac-41d7-a37a-6455bafd4735"
/>
<br/><br/>

This uses [Brussels](https://github.com/oxidecomputer/brussels), the new
SP and RoT release tooling I'm working on, and in particular the changes
I introduced in oxidecomputer/brussels#3. Brussels is fetched as a
pre-built binary from its GitHub Release to ensure it doesn't delay the
release process by trying to compile it.

The `brussels init` command will be executed in its own job as the first
build step. It will look at all the existing tags to determine what is
the next version number, calculate the tag name, and expose all of that
information into `GITHUB_OUTPUT` to let the rest of the workflow know
the version to use. It will also write all the metadata it gathered in a
`brussels-manifest.json` file, so that future Brussels steps can
reference that metadata.

The tag will not be created and pushed by `brussels init`. Instead, we
will let the "create GitHub Release" step at the end of the workflow
create the tag on its own, as GitHub will do so when creating a release
without an underlying tag.

While I was there, I also changed the build process to inject the
version number directly during the build, instead of waiting for
permslip to add it as part of the signature process.

This is just the first improvement to the SP release process that I'm
going to do. The end goal is for SP releases to be fully automated
(either by requiring a single button click or by putting the whole
release process in a cronjob).

Note that by default the workflow will use the latest release of
Brussels. An optional field when starting the workflow run lets you
download Brussels from a CI job in the Brussels repository. This mostly
exists to help me test changes without publishing endless releases.