psu_seq: refactor ereports to use `microcbor`
PR #2246 introduced the `microcbor` crate for derivable CBOR
serialization with a fixed max length. That branch changed the ereports
in `drv_gimlet_seq_server` to use the new derives. The intent at the
time was to update other existing ereports, such as those in
`drv_psc_seq_server`, separately. This branch does that.