Only read DIMMs in A0HP (#2287)
This includes a workaround for
https://github.com/oxidecomputer/quartz/issues/434; 

and has FPGA fixes for https://github.com/oxidecomputer/quartz/issues/437 and https://github.com/oxidecomputer/quartz/issues/435 which resulted in topo not showing the correct number of dimms or the correct information for the dimms.

I've cycled this ~100 boots without seeing any dimm training failures  so this seems to be reliable.
---------

Co-authored-by: Nathanael Huffman <nathanael@oxidecomputer.com>