Update to Brussels 0.2.0 (#2300)
This PR updates Brussels to 0.2.0 (oxidecomputer/brussels#7) and makes
the necessary changes to be compatible with it.

This release of Brussels doesn't change the manual steps needed to
publish a release (it still only does the tagging and initial build),
and instead contains these changes:

* Records the commit we are releasing in the manifest, so that the rest
of the (future) release process will be able to rely on it.
* Rather than hardcoding the release groups in the Brussels source code,
there is now a file in the Hubris repository, `brussels.toml`. The fine
defines both how to create the tag and the list of boards to release.

Along with the update I changed the GitHub release creation code to use
the commit we are actually releasing, instead of releasing the latest
commit on `main` unconditionally. This will prevent cases like the
all-sp-v1.0.49 tag pointing to the wrong commit.