Merge PR workflows into a single one (#2301)
This PR merges the four workflows that run in every PR into a single
one, and adds a "CI finished" job that executes when all workflows
finish. This has a few consequences:

* Once we gate PRs on the new "CI finished" job instead of individual
jobs, all jobs will be required for a PR to be merged. Adding new
required jobs or renaming existing ones won't need admin access to
change branch protection anymore.

* Cancelling CI for a PR will require cancelling only one workflow
rather than four.

* The "Actions" tab will be cleaner, with only one execution per PR and
less workflows to scan through on the sidebar.

I remember talking with folks a few weeks/months ago about this, and
there was a concern about having to wait more time before being able to
click the merge button if we make all of the workflows required. IMO
this is less of a concern now that we can mark a PR as "merge when all
required checks pass", does that alleviate the concerns?