net-api: change TryFrom<IpAddress> to just From
It turns out, the TryFrom impl for smoltcp's IpAddress can't fail.
This changed in 2023 with the upgrade to smoltcp 0.9, which made
the shape of IpAddress depend on which protocol features we'd
selected, removing the "unreachable" variants.

This commit converts the impl to From, making it clear that it can't
fail and eliminating an unwrap at its (apparently single) use site.