README: change CI platform language
I realized that this section of the README was sounding like an
invitation for potential starter projects, without giving people full
context on what the "maintain" part of maintaining a tier-one platform
would involve (long term commitment from an experienced engineer). In
practice, macOS seems to be the main "other" platform that folks are
interested in, but it's also the only one that those of us without Apple
hardware can't test locally -- every other OS permits virtualization,
which makes macOS a uniquely high support burden.

To avoid misleading people, I've adjusted the language to remove the
suggestion that we're looking for maintainers. It's not that we _aren't_
looking for maintainers, but more that maintaining a platform will be a
longer discussion and not merely a PR.